Connection Code cleanup and better interfaces
trying to make interfaces clean for multiconnection patch
